Moved the tests EntityHelper into separate folder as it was clashing with the folder name that would be used for actual main project's Helper tests
