+++ cv19ResSupportV3.Tests/cv19ResSupportV3.Tests.csproj
@@ -35,4 +35,7 @@
     </Content>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="V4\TestHelpers\" />
+  </ItemGroup>
 </Project>
